Split index module
