+++ src/index/mod.rs
@@ -1,0 +1,5 @@
+#[allow(clippy::module_inception)]
+mod index;
+mod range;
+
+pub use self::{index::Index, range::Range};
